Locations references to pointers
location.neighbors --> std::map<std::string, Location*> neighbors;
this means any access to a location needs to be dereferenced.

I had to create the locations in the constructor, and now create_world() because if you push_back the location in the create_world() it creates a copy. the memory of the map is lost. So you create the locations, then call create_world and it makes everything else and maps everything out.

game.locations --> std::vector<std::reference_wrapper<Location>>
this holds reference wrappers because the locations are already stored in the game class. And the rest of the code was already written to work with location objects and not pointer objects.

game.current_location is also a pointer that needs to be derefenced when its used. this is a pointer because we don't want to create copies.